raw-converter/requirements.txt
PyExifTool==0.1.1